refactor: [GSW-2233] RouterFee util
